fix studio and search navigation
